Added async tests, moved tests to reader.rs
