+++ parquet/src/file/metadata/dictionary.rs
@@ -154,108 +154,3 @@ pub fn is_utf8_column(column_descr: &ColumnDescriptor) -> bool {
         ConvertedType::UTF8 | ConvertedType::JSON
     )
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::file::metadata::ParquetMetaDataReader;
-    use crate::util::test_common::file_util::get_test_file;
-    use arrow_array::{Array, BinaryArray, StringArray};
-    use bytes::Bytes;
-    use std::io::Read;
-
-    /// Helper to read a test file into bytes
-    fn read_test_file(file_name: &str) -> Bytes {
-        let mut file = get_test_file(file_name);
-        let mut buffer = Vec::new();
-        file.read_to_end(&mut buffer).unwrap();
-        Bytes::from(buffer)
-    }
-
-    #[test]
-    fn test_read_column_dictionary_uncompressed() {
-        // plain-dict-uncompressed-checksum.parquet: col 1 = binary_field (BYTE_ARRAY dict, uncompressed)
-        // Note: No STRING annotation, so returns BinaryArray
-        let file_data = read_test_file("plain-dict-uncompressed-checksum.parquet");
-        let reader = ParquetMetaDataReader::new();
-        let metadata = reader.parse_and_finish(&file_data).unwrap();
-
-        let dict = ParquetMetaDataReader::read_column_dictionary(&file_data, &metadata, 0, 1).unwrap();
-
-        assert!(dict.is_some());
-        let dict_array = dict.unwrap();
-        let binary_array = dict_array
-            .as_any()
-            .downcast_ref::<BinaryArray>()
-            .expect("Expected BinaryArray");
-        assert!(binary_array.len() > 0);
-    }
-
-    #[test]
-    fn test_read_column_dictionary_binary() {
-        // alltypes_dictionary.parquet: col 8 = date_string_col (BYTE_ARRAY with dict, NO STRING annotation)
-        let file_data = read_test_file("alltypes_dictionary.parquet");
-        let reader = ParquetMetaDataReader::new();
-        let metadata = reader.parse_and_finish(&file_data).unwrap();
-
-        let dict = ParquetMetaDataReader::read_column_dictionary(&file_data, &metadata, 0, 8).unwrap();
-
-        assert!(dict.is_some());
-        let dict_array = dict.unwrap();
-        let binary_array = dict_array
-            .as_any()
-            .downcast_ref::<BinaryArray>()
-            .expect("Expected BinaryArray for pure BYTE_ARRAY column");
-        assert!(binary_array.len() > 0);
-    }
-
-    #[test]
-    fn test_read_column_dictionary_none() {
-        // alltypes_tiny_pages_plain.parquet: col 9 = string_col (PLAIN encoding, NO dictionary)
-        let file_data = read_test_file("alltypes_tiny_pages_plain.parquet");
-        let reader = ParquetMetaDataReader::new();
-        let metadata = reader.parse_and_finish(&file_data).unwrap();
-
-        let dict = ParquetMetaDataReader::read_column_dictionary(&file_data, &metadata, 0, 9).unwrap();
-
-        assert!(dict.is_none());
-    }
-
-    #[test]
-    fn test_read_column_dictionary_compressed() {
-        // rle-dict-snappy-checksum.parquet: col 1 = binary_field (BYTE_ARRAY with Snappy-compressed dict)
-        let file_data = read_test_file("rle-dict-snappy-checksum.parquet");
-        let reader = ParquetMetaDataReader::new();
-        let metadata = reader.parse_and_finish(&file_data).unwrap();
-
-        // Verify compression
-        assert_eq!(
-            metadata.row_group(0).column(1).compression(),
-            crate::basic::Compression::SNAPPY
-        );
-
-        let dict = ParquetMetaDataReader::read_column_dictionary(&file_data, &metadata, 0, 1).unwrap();
-
-        assert!(dict.is_some());
-        let dict_array = dict.unwrap();
-        // binary_field has no STRING annotation, so expect BinaryArray
-        let binary_array = dict_array
-            .as_any()
-            .downcast_ref::<BinaryArray>()
-            .expect("Expected BinaryArray");
-        assert!(binary_array.len() > 0);
-    }
-
-    #[test]
-    fn test_read_column_dictionary_wrong_type() {
-        // plain-dict-uncompressed-checksum.parquet: col 0 = long_field (INT32 with dict)
-        let file_data = read_test_file("plain-dict-uncompressed-checksum.parquet");
-        let reader = ParquetMetaDataReader::new();
-        let metadata = reader.parse_and_finish(&file_data).unwrap();
-
-        // INT32 column should fail
-        let result = ParquetMetaDataReader::read_column_dictionary(&file_data, &metadata, 0, 0);
-
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("BYTE_ARRAY"));
-    }
-}
